+++ client/webapi.go
@@ -8,195 +8,19 @@ import (
 	"io"
 	"log"
 	"net/http"
-	"time"
 
 	"github.com/ipfs/go-cid"
 )
 
 const isAssetAlreadyExist = 1017
 
-type JWTPayload struct {
-	// role base access controller permission
-	Allow []string
-	ID    string
-	// TODO remove NodeID later, any role id replace as ID
-	NodeID string
-	// Extend is json string
-	Extend string
-	// The sub permission of user
-	AccessControlList []string
-}
-
-// AssetProperty represents the properties of an asset.
-type AssetProperty struct {
-	AssetCID  string
-	AssetName string
-	AssetSize int64
-	AssetType string
-	NodeID    string
-	GroupID   int
-}
-
-type CreateAssetReq struct {
-	// UserID string
-	AreaID string
-	AssetProperty
-}
-
-// CreateAssetRsp represents the response when creating an asset.
-type Endpoint struct {
-	CandidateAddr string
-	Token         string
-	AlreadyExists bool
-}
-
-type CreateAssetRsp struct {
-	IsAlreadyExist bool
-	Endpoints      []*Endpoint
-}
-
-// CandidateIPInfo represents information about a candidate IP.
-type CandidateIPInfo struct {
-	NodeID      string
-	IP          string
-	ExternalURL string
-}
-
-// ReplicaInfo represents information about a replica.
-type ReplicaInfo struct {
-	Hash        string
-	NodeID      string
-	Status      int
-	IsCandidate bool
-	EndTime     time.Time
-	DoneSize    int64
-}
-
-// AssetRecord represents information about an asset record.
-type AssetRecord struct {
-	CID                   string
-	Hash                  string
-	NeedEdgeReplica       int64
-	TotalSize             int64
-	TotalBlocks           int64
-	Expiration            time.Time
-	CreatedTime           time.Time
-	EndTime               time.Time
-	NeedCandidateReplicas int64
-	ServerID              string
-	State                 string
-	NeedBandwidth         int64
-
-	RetryCount        int64
-	ReplenishReplicas int64
-	ReplicaInfos      []*ReplicaInfo
-
-	SPCount int64
-}
-
-// UserAssetDetail represents detailed information about a user's asset.
-type UserAssetDetail struct {
-	UserID      string
-	Hash        string
-	AssetName   string
-	AssetType   string
-	ShareStatus int64
-	Expiration  time.Time
-	CreatedTime time.Time
-	TotalSize   int64
-}
-
-// AssetOverview represents an overview of an asset.
-type AssetOverview struct {
-	AssetRecord      *AssetRecord
-	UserAssetDetail  *UserAssetDetail
-	VisitCount       int
-	RemainVisitCount int
-}
-
-// ListAssetRecordRsp represents the response when listing asset records.
-type ListAssetRecordRsp struct {
-	Total          int
-	AssetOverviews []*AssetOverview
-}
-
-// AssetGroup user asset group
-type AssetGroup struct {
-	ID          int
-	UserID      string
-	Name        string
-	Parent      int
-	AssetCount  int
-	AssetSize   int64
-	CreatedTime time.Time
-}
-
-// ListAssetGroupRsp list  asset group records
-type ListAssetGroupRsp struct {
-	Total       int           `json:"total"`
-	AssetGroups []*AssetGroup `json:"list"`
-}
-
-// UserAssetSummary user asset and group
-type UserAssetSummary struct {
-	AssetOverview *AssetOverview
-	AssetGroup    *AssetGroup
-}
-
-// ListAssetSummaryRsp list asset and group
-type ListAssetSummaryRsp struct {
-	Total int                 `json:"total"`
-	List  []*UserAssetSummary `json:"list"`
-}
-
-type UploadInfo struct {
-	List          []*NodeUploadInfo
-	AlreadyExists bool
-	AreaID        string
-	Log           string
-}
-
-type NodeUploadInfo struct {
-	UploadURL string
-	Token     string
-	NodeID    string
-}
-
-type VipInfo struct {
-	UserID string `json:"uid"`
-	VIP    bool   `json:"vip"`
-}
-
-type ShareAssetResult struct {
-	AssetCID string   `json:"asset_cid"`
-	Redirect bool     `json:"redirect"`
-	Size     int64    `json:"size"`
-	URLs     []string `json:"url"`
-	FileName string
-}
-
-type Result struct {
-	Code int    `json:"code"`
-	Err  int    `json:"err"`
-	Msg  string `json:"msg"`
-	Data interface{}
-}
-
 const (
 	AssetTransferTypeUpload   = "upload"
 	AssetTransferTypeDownload = "download"
-)
 
-type AssetTransferReq struct {
-	UserId       string `json:"user_id"`
-	Cid          string `json:"cid"`
-	Hash         string `json:"hash"`
-	Rate         int64  `json:"rate"`
-	CostMs       int64  `json:"cost_ms"`
-	TotalSize    int64  `json:"total_size"`
-	Succeed      bool   `json:"succeed"`
-	TransferType string `json:"transfer_type"`
-}
+	AssetTransferStateSuccess = 1
+	AssetTransferStateFailed  = 2
+)
 
 // Webserver defines the interface for the scheduler.
 type Webserver interface {
@@ -344,6 +168,7 @@ type webCreateAssetReq struct {
 	AssetSize int64    `json:"asset_size"`
 	GroupID   int64    `json:"group_id"`
 	Encrypted bool     `json:"encrypted"`
+	NeedTrace bool     `json:"need_trace"`
 }
 
 // CreateUserAsset creates a new user asset.
@@ -714,7 +539,7 @@ func (s *webserver) GetAPPKeyPermissions(ctx context.Context, userID, keyName st
 
 // GetNodeUploadInfo
 func (s *webserver) GetNodeUploadInfo(ctx context.Context, userID string, urlMode bool) (*UploadInfo, error) {
-	url := fmt.Sprintf("%s/api/v1/storage/get_upload_info?encrypted=false&area_id=%s", s.url, s.area)
+	url := fmt.Sprintf("%s/api/v1/storage/get_upload_info?encrypted=false&area_id=%s&need_trace=true", s.url, s.area)
 	if urlMode {
 		url += "&urlMode=true"
 	}
@@ -784,7 +609,7 @@ func (s *webserver) AssetTransferReport(ctx context.Context, req AssetTransferRe
 	// 	TransferType: transferType,
 	// }
 
-	if req.Succeed {
+	if req.State == AssetTransferStateSuccess {
 		// bytes per second
 		req.Rate = req.TotalSize / req.CostMs * 1000
 	}
